remove template interfaces from augur-ui
