+++ packages/augur-ui/src/modules/types.ts
@@ -729,32 +729,6 @@ export interface SortedGroup {
   autoCompleteList?: Array<SortedGroup>;
 }
 
-
-export interface UserInputText {
-  value: string;
-}
-
-export interface UserInputDateYear extends UserInputText {}
-export interface UserInputDateTime {
-  endTime: number;
-  hour?: number;
-  minute?: number;
-  meridiem: string;
-  timezone: string;
-  offset: number;
-  offsetName: string;
-  endTimeFormatted: TimezoneDateObject;
-}
-export interface UserInputDropdown extends UserInputText {}
-export interface UserInputUserOutcome extends UserInputText {}
-
-export type UserInputtedType =
-  | UserInputText
-  | UserInputDateYear
-  | UserInputDateTime
-  | UserInputDropdown
-  | UserInputUserOutcome;
-
 export interface CategoryList {
   [category: string]: [
     {
@@ -765,53 +739,5 @@ export interface CategoryList {
       ];
     }
   ];
-}
-export interface TemplateChildren {
-  [category: string]: CategoryTemplate;
-}
-
-export interface CategoryTemplate {
-  templates: Template[];
-  children: TemplateChildren;
-}
-
-export interface ResolutionRule {
-  text: string;
-  isSelected?: boolean;
-}
-
-export interface ResolutionRules {
-  [REQUIRED]?: ResolutionRule[];
-  [CHOICE]?: ResolutionRule[];
-}
-
-export interface Template {
-  hash: string;
-  categories: Categories;
-  marketType: string;
-  question: string;
-  example: string;
-  inputs: TemplateInput[];
-  resolutionRules: ResolutionRules;
-  denomination?: string;
-  tickSize?: number;
-}
-
-export interface TemplateInput {
-  id: number;
-  type: TemplateInputType;
-  placeholder: string;
-  label?: string;
-  tooltip?: string;
-  userInput?: string;
-  userInputObject?: UserInputtedType;
-  values?: ValueLabelPair[];
-  sublabel?: string;
-}
-
-export interface Categories {
-  primary: string;
-  secondary: string;
-  tertiary: string;
 }
